requirements.txt
absl-py==2.2.2
anyascii==0.3.2
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
defusedxml==0.7.1
editdistance==0.8.1
efficientnet==1.0.0
essential-generators==1.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
imgaug==0.4.0
Jinja2==3.1.6
keras==3.9.2
Keras-Applications==1.0.8
keras-ocr==0.9.3
kiwisolver==1.4.8
langdetect==1.0.9
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
ninja==1.11.1.4
numpy==1.26.4
onnx==1.17.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pillow==11.2.1
protobuf==4.25.6
pyclipper==1.3.0.post6
Pygments==2.19.1
pyparsing==3.2.3
pypdfium2==4.30.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-doctr==0.11.0
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.3
rich==14.0.0
scikit-image==0.25.2
scipy==1.15.2
setuptools==78.1.0
shapely==2.1.0
six==1.17.0
sympy==1.13.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==3.0.1
tifffile==2025.3.30
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
validators==0.34.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
